(tf-gcloud): Do not launch browser when logging in
Started getting some weird error from interactive login commands today:

[237457:237457:0828/100833.755856:ERROR:components/sharing_message/sharing_service.cc:297] Device registration failed with fatal error
[237499:237499:0828/100839.962848:ERROR:gpu/command_buffer/service/gles2_cmd_decoder_passthrough.cc:1096] [GroupMarkerNotSet(crbug.com/242999)/home/sergio/Projects/personal/devex-makefile/git st0301E00AC0C0000]Automatic fallback to software WebGL has been deprecated. Please use the --enable-unsafe-swiftshader (about:flags#enable-unsafe-swiftshader) flag to opt in to lower security guarantees for trusted content.

Enabling graphics acceleration in brave got rid of the second error,
but the first one is still there.

In addition, it looks like the browser still runs in the background
after the login is complete even if I close the tab, and no other tabs are open,
which is a bit annoying.

Seems like using --no-launch-browser by default is a safer option to get
rid of these issues.